requirements.txt
uvicorn==0.32.0
fastapi==0.115.3
loguru==0.7.2
python-dotenv==1.0.1
pyjwt==2.9.0
loguru==0.7.2
email-validator==2.2.0
asyncpg==0.30.0
passlib==1.7.4
bcrypt==4.3.0
sqlalchemy==2.0.39
psycopg2==2.9.10
